# apps/api/specs/013-performance-caching-observability-operations/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Masarifi SPEC-BE-013 Operations API
  version: 1.0.0
  description: Bounded Free-only operations, configuration, performance, recovery, and safe client metadata contracts.
servers:
  - url: /
security:
  - ClerkBearer: []
paths:
...
  /api/v1/meta:
    get:
      operationId: getPlatformMetadata
      tags: [Client configuration]
      parameters:
        - name: X-Masarifi-Platform
          in: header
          schema: { type: string, enum: [ios, android, admin] }
        - name: X-Masarifi-App-Version
          in: header
          schema: { type: string, maxLength: 32 }
        - name: X-Masarifi-Locale
          in: header
          schema: { type: string, enum: [ar, en] }
      responses:
        '200':
          description: Safe Free-only metadata
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, max-age=30' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformMetadata' }
        '400': { $ref: '#/components/responses/SafeError' }
        '401': { $ref: '#/components/responses/SafeError' }
        '503': { $ref: '#/components/responses/SafeError' }
components:
  securitySchemes:
    ClerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    SafeError:
      description: Stable safe error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    OperationsTimestamp:
      type: string
      format: date-time
...
    ApiError:
      type: object
      additionalProperties: false
      required: [statusCode, code, message, requestId]
      properties:
        statusCode: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        requestId: { type: string, minLength: 1, maxLength: 128 }
        fieldErrors:
          type: object
          maxProperties: 50
          additionalProperties:
            type: array
            maxItems: 10
            items: { type: string, maxLength: 120 }
...
    LocalizedMessage:
      type: object
      additionalProperties: false
      required: [ar, en]
      properties:
        ar: { type: string, minLength: 1, maxLength: 240 }
        en: { type: string, minLength: 1, maxLength: 240 }
...
    PlatformCapabilities:
      type: object
      additionalProperties: false
      required: [coreFinanceAvailable, billingAvailable, paidEntitlement, checkoutAvailable, subscriptionManagementAvailable, promotionsAvailable, aiAvailable, aiAllowancePerRolling24Hours]
      properties:
        coreFinanceAvailable: { type: boolean, const: true }
        billingAvailable: { type: boolean, const: false }
        paidEntitlement: { type: boolean, const: false }
        checkoutAvailable: { type: boolean, const: false }
        subscriptionManagementAvailable: { type: boolean, const: false }
        promotionsAvailable: { type: boolean, const: false }
        aiAvailable: { type: boolean }
        aiAllowancePerRolling24Hours: { type: integer, const: 5 }
    SafeMaintenance:
      type: object
      additionalProperties: false
      required: [active, scopes, message]
      properties:
        active: { type: boolean }
        scopes: { type: array, maxItems: 10, uniqueItems: true, items: { type: string } }
        message: { anyOf: [{ $ref: '#/components/schemas/LocalizedMessage' }, { type: 'null' }] }
    PlatformMetadata:
      type: object
      additionalProperties: false
      required: [apiVersion, serverTime, minMobileVersion, minAdminVersion, capabilities, maintenance, featureFlags]
      properties:
        apiVersion: { type: string, const: v1 }
        serverTime: { $ref: '#/components/schemas/OperationsTimestamp' }
        minMobileVersion: { type: [string, 'null'], maxLength: 32 }
        minAdminVersion: { type: [string, 'null'], maxLength: 32 }
        capabilities: { $ref: '#/components/schemas/PlatformCapabilities' }
        maintenance: { $ref: '#/components/schemas/SafeMaintenance' }
        featureFlags:
          type: object
          additionalProperties:
            type: boolean
          maxProperties: 50
        configurationVersion: { type: integer, minimum: 1 }
...